refactor: extract CslTenantCheck from CslAuthorizationCheck
CslAuthorizationCheck bundled four concerns in one class: principal
skip-logic, RBAC checks against the CSL AuthorizationCheckPort,
tenant-membership checks against AuthorizedTenantsResolver, and
combined RBAC+tenant convenience methods. Tenant logic never touches
the authz port, which made it a clean seam to split out and kept the
authz class from growing into a god class as more checks land on it.

Move resolveAuthorizedTenants, isMultiTenancyChecksEnabled,
checkTenant, and checkTenantsRequiringPrincipal verbatim into a new
CslTenantCheck, constructed from only TokenClaimsAuthenticationResolver
and EngineSecurityConfig. CslAuthorizationCheck keeps its existing
constructor signature and all four public tenant methods as thin
delegates to an internally-held CslTenantCheck, so none of the ~14
existing call sites or the two DI construction sites
(EngineProcessors, MessageEventProcessors) need to change. Javadoc
cross-references were re-qualified to point at the class each method
now lives on, without altering the wording of their behavioral
contracts.

Split CslAuthorizationCheckTest's tenant-related tests into a new
CslTenantCheckTest, constructing CslTenantCheck directly; the thin
delegates on CslAuthorizationCheck need no new tests of their own.
